making it a volume estimation example
